requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.2.10=py_0
astroid=2.4.2=py37_0
atomicwrites=1.4.0=py_0
attrs=19.3.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
ca-certificates=2020.6.24=0
certifi=2020.6.20=py37_0
chardet=3.0.4=pypi_0
click=7.1.2=py_0
colorama=0.4.3=py_0
configparser=5.0.0=pypi_0
decorator=4.4.2=py_0
django=3.0.3=py_0
docker-pycreds=0.4.0=pypi_0
flask=1.1.2=py_0
gitdb=4.0.5=pypi_0
gitpython=3.1.7=pypi_0
gql=0.2.0=pypi_0
graphql-core=1.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.10=pypi_0
importlib-metadata=1.7.0=py37_0
importlib_metadata=1.7.0=0
iniconfig=1.0.1=py_0
intel-openmp=2019.4=245
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.17.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
isort=4.3.21=py37_0
itsdangerous=1.1.0=py37_0
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
joblib=0.16.0=py_0
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py37_0
lazy-object-proxy=1.4.3=py37he774522_0
libsodium=1.0.18=h62dcd97_0
markupsafe=1.1.1=py37hfa6e2cd_1
mccabe=0.6.1=py37_1
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.14=py37h6288b17_0
mkl_random=1.0.4=py37h343c172_0
more-itertools=8.4.0=py_0
numpy=1.17.0=py37h19fb1c0_0
numpy-base=1.17.0=py37hc3f5095_0
nvidia-ml-py3=7.352.0=pypi_0
openssl=1.1.1g=he774522_0
packaging=20.4=py_0
pandas=1.1.0=py37ha925a31_0
parso=0.7.0=py_0
pathtools=0.1.2=pypi_0
pickleshare=0.7.5=py37_1001
pip=20.2.1=py37_0
pluggy=0.13.1=py37_0
promise=2.3=pypi_0
prompt-toolkit=3.0.5=py_0
psutil=5.7.2=pypi_0
py=1.9.0=py_0
pygments=2.6.1=py_0
pylint=2.5.3=py37_0
pyparsing=2.4.7=py_0
pytest=6.0.1=py37_0
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=py_0
pywin32=227=py37he774522_1
pyyaml=5.3.1=pypi_0
pyzmq=19.0.1=py37ha925a31_1
requests=2.24.0=pypi_0
scikit-surprise=1.1.1=py37h44b1f71_0
sentry-sdk=0.16.4=pypi_0
setuptools=49.2.1=py37_0
shortuuid=1.0.1=pypi_0
six=1.15.0=py_0
smmap=3.0.4=pypi_0
sqlite=3.32.3=h2a8f88b_0
sqlparse=0.3.1=py_0
subprocess32=3.5.4=pypi_0
toml=0.10.1=py_0
tornado=6.0.4=py37he774522_1
traitlets=4.3.3=py37_0
typed-ast=1.4.1=py37he774522_0
urllib3=1.25.10=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wandb=0.9.4=pypi_0
watchdog=0.10.3=pypi_0
wcwidth=0.2.5=py_0
werkzeug=1.0.1=py_0
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
zeromq=4.3.2=ha925a31_2
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
